tests/integration: Powerc - Add fixup section test
Add testcase to test fixup section changes on Powerpc. Currently we
do not support such changes and attempting to build the livepatch
module using this testcase should fail until we start supporting the
fixup section changes. Also, append '-FAIL' to the name of the
testcase hinting kpatch-test that this testcase should fail as if now.

Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>